Refactor `__init__.py` to simplify imports by restructuring and reducing redundancy.
