+++ neuro_admin_client/entities.py
@@ -6,6 +6,31 @@ from typing import Any, List, Optional, Union, cast
 
 from yarl import URL
 
+__all__ = [
+    "FullNameMixin",
+    "UserInfo",
+    "User",
+    "Balance",
+    "Quota",
+    "ClusterUserRoleType",
+    "Cluster",
+    "OrgNotificationIntervals",
+    "Org",
+    "OrgUserRoleType",
+    "OrgUser",
+    "OrgUserWithInfo",
+    "OrgCluster",
+    "ClusterUser",
+    "ClusterUserWithInfo",
+    "ProjectUserRoleType",
+    "Project",
+    "ProjectUser",
+    "ProjectUserWithInfo",
+    "OrderedEnum",
+    "Action",
+    "Permission",
+]
+
 
 class FullNameMixin:
     first_name: Optional[str]
